doomday

+++ src/main/scala/doomsday/core/package.scala
@@ -4,5 +4,5 @@ import breeze.linalg.DenseMatrix
 
 package object core {
   def as_var(x: Int) : Var = new Var(DenseMatrix(x.toDouble))
-  def as_var(x: Double) : Var = new Var(DenseMatrix(x.toDouble))
+  def as_var(x: Double) : Var = new Var(DenseMatrix(x))
 }
